8198895: Compilation errors in java.desktop with VS 2017

+++ src/java.desktop/windows/native/libawt/windows/awt_Toolkit.cpp
@@ -1106,7 +1106,7 @@ LRESULT CALLBACK AwtToolkit::WndProc(HWND hWnd, UINT message,
           if (comp != NULL)
           {
               comp->SetInputMethod(self, useNativeCompWindow);
-              comp->ImmAssociateContext((HIMC)context);
+              comp->ImmAssociateContext((HIMC)((intptr_t)context));
           }
 
           if (peer != NULL) {
@@ -3139,7 +3139,7 @@ void AwtToolkit::ShowTouchKeyboard() {
         (m_touchKbrdExeFilePath != NULL)) {
         HINSTANCE retVal = ::ShellExecute(NULL, _T("open"),
             m_touchKbrdExeFilePath, NULL, NULL, SW_SHOW);
-        if ((int)retVal <= 32) {
+        if ((int)((intptr_t)retVal) <= 32) {
             DTRACE_PRINTLN1("AwtToolkit::ShowTouchKeyboard: Failed"
                 ", retVal='%d'", (int)retVal);
         }
